1：新增用户地址相关实体类 2：新增用户地址相关CRUD 3：新增上下文工具类Context和当前登录用户类Current

+++ src/main/java/com/xiaoteng/server/controller/user/UserController.java
@@ -40,7 +40,7 @@ public class UserController {
      * 用户查询功能的实现
      */
     @GetMapping("/query")
-    public Result<SysUser>  queryUserById(@RequestParam("id") Integer id)
+    public Result<SysUser>  queryUserById(@RequestParam("id") Long id)
     {
         SysUser user = userServiceImpl.queryUserById(id);
